### 2018-08-01 #810 @stenskjaer
I'm a little bit uncertain whether I am expected to react to this, or I just wait until reviewers are assigned?

I could maybe make some minor changes to the text, now that I have seen the PDF proof. But I'm not sure this is the time and place.
