### 2018-08-02 #810 @arfon
> I'm a little bit uncertain whether I am expected to react to this, or I just wait until reviewers are assigned?

Hi @stenskjaer - we're currently waiting for an editor to pick up this submission. We currently have close to 100 submissions in progress and so are a little behind. Thanks for your patience.

In the meantime, if you want to edit your paper, please do so and then as @whedon to recompile it with:
