### 2018-08-05 88d09af
Create minor updates of wording
