### 2018-08-28 be94eb2
Update the regex version requirement.
