### 2018-08-28 5cb1618
Correct faulty use of temporary settings in tests.
