+++ samewords/test/test_matcher.py
@@ -1256,13 +1256,13 @@ class TestDefineSearchWords:
         assert search_words == expect
 
     def test_custom_ellipses_with_space(self):
-        old_exclude = settings['exclude_macros']
-        settings['exclude_macros'] += [
+        excluded = settings['exclude_macros'] + [
             "\\.\\.\\.",
             "-+",
             "\\,-\\,",
             "\\\\,-+\\\\,"
         ]
+        expect = ['A', 'E']
         single_dash = r'\edtext{A B C D E}{\lemma{A - E}\Afootnote{}}'
         double_dash = r'\edtext{A B C D E}{\lemma{A -- E}\Afootnote{}}'
         triple_dash = r'\edtext{A B C D E}{\lemma{A --- E}\Afootnote{}}'
@@ -1274,19 +1274,18 @@ class TestDefineSearchWords:
         dots_brackets = r'\edtext{A B C D E}{\lemma{A \dots{} E}\Afootnote{}}'
         ldots = r'\edtext{A B C D E}{\lemma{A \ldots E}\Afootnote{}}'
         ldots_brackets = r'\edtext{A B C D E}{\lemma{A \ldots{} E}\Afootnote{}}'
-        expect = ['A', 'E']
-        assert self.run_wordlist(single_dash) == expect
-        assert self.run_wordlist(double_dash) == expect
-        assert self.run_wordlist(triple_dash) == expect
-        assert self.run_wordlist(endash) == expect
-        assert self.run_wordlist(emdash) == expect
-        assert self.run_wordlist(comma_string) == expect
-        assert self.run_wordlist(thin_space) == expect
-        assert self.run_wordlist(dots) == expect
-        assert self.run_wordlist(dots_brackets) == expect
-        assert self.run_wordlist(ldots) == expect
-        assert self.run_wordlist(ldots_brackets) == expect
-        settings['exclude_macros'] = old_exclude
+        with temp_settings({'exclude_macros': excluded}):
+            assert self.run_wordlist(single_dash) == expect
+            assert self.run_wordlist(double_dash) == expect
+            assert self.run_wordlist(triple_dash) == expect
+            assert self.run_wordlist(endash) == expect
+            assert self.run_wordlist(emdash) == expect
+            assert self.run_wordlist(comma_string) == expect
+            assert self.run_wordlist(thin_space) == expect
+            assert self.run_wordlist(dots) == expect
+            assert self.run_wordlist(dots_brackets) == expect
+            assert self.run_wordlist(ldots) == expect
+            assert self.run_wordlist(ldots_brackets) == expect
 
     def test_custom_ellipses_without_space(self):
         single_dash = r'\edtext{A B C D E}{\lemma{A-E}\Afootnote{}}'
