### 2018-08-28 b6ca539
Pop version patch number.
