### 2018-08-28 f7c1cbf
Split document and string processing into two.
This makes the internal API more flexible.
