### 2018-08-28 e376d9f
Tokenize latex star commands correctly
