+++ samewords/test/test_core.py
@@ -6,11 +6,11 @@ import samewords.document as document
 class TestMainProcessing:
     unprocessed = './samewords/test/assets/da-49-l1q1.tex'
     processed = './samewords/test/assets/da-49-l1q1-processed.tex'
-    unprocessed_content = document.doc_content(unprocessed)
-    processed_content = document.doc_content(processed)
+    unproc_content = document.doc_content(unprocessed)
+    proc_content = document.doc_content(processed)
 
     def test_process_document(self):
-        assert process_document(self.unprocessed) == self.processed_content
+        assert process_document(self.unprocessed) == self.proc_content
 
     def test_update_document(self):
         unupdated = './samewords/test/assets/simple-unupdated.tex'
@@ -18,6 +18,8 @@ class TestMainProcessing:
         updated_content = document.doc_content(updated)
         assert process_document(unupdated, 'update') == updated_content
 
-
     def test_clean_document(self):
-        assert process_document(self.processed, 'clean') == self.unprocessed_content
+        assert process_document(self.processed, 'clean') == self.unproc_content
+
+    def test_process_string(self):
+        assert process_string(self.unproc_content) == self.proc_content
