### 2018-08-28 9bc0a05
Tokenize edtext after word content as separate word
