### 2018-08-28 59d2a60
Case insensitive in edtext index matching
