### 2018-08-28 1d5317c
Update the version and correct changelog
