### 2018-08-29 #810 @Kevin-Mattheus-Moerman
@stenskjaer thanks for your submission. From reading your paper it seems your work is a useful package to automate the disambiguation of identical words e.g. when writing critical notes. It is unclear to me however how this work is relevant to research (I am not an expert in this field). The [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) state:

> The software should have an **obvious** research application

Could you please have a read through the submission requirements and ensure that your paper clearly discusses research applications, and how the software influences research outcomes?
