### 2018-09-07 #810 @stenskjaer
@Kevin-Mattheus-Moerman thank you for your response.

I think it is important to stress that the creation of critical text editions and textual criticism is the most foundational research you can find when it comes to any text interpretative discipline. The purpose of the activity is to provide texts that will be the basis for any subsequent work, be it analysis by a researcher, translation for further sedimentation, or leisurely reading. This activity thus provides the raw material for any kind of research within a given discipline.

Establishing such texts, be it in English, German, French or the classical languages such as Greek and Latin in my own case, is a highly specialized activity presupposing research skills in all the relevant subject areas of a text, such as the linguistic disciplines, the study of the content, but also the whole discipline of stemmatics, codicology, palaeography (it very often involves reading medieval or early modern material in manuscript form), and text history. For some descriptions of these different branches of the discipline, see [the wikipedia article](https://en.wikipedia.org/wiki/Textual_criticism).

The [review and vetting sheet for the Modern Language Association](https://www.mla.org/Resources/Research/Surveys-Reports-and-Other-Documents/Publishing-and-Scholarship/Reports-from-the-MLA-Committee-on-Scholarly-Editions/Guidelines-for-Editors-of-Scholarly-Editions) for critical text editions will give you an impression of some of the requirements of a critical text edition.

To get an impression of what the results of such textual work looks like, I can point you to one of my own publications: https://cimagl.saxo.ku.dk/download/86/86christensen79-131.pdf.
But you can also have a look at this picture of a page from the standard critical edition of the New Testament in Greek.
![c5euc84lap8jf1dnr](https://user-images.githubusercontent.com/4047523/45203924-fa03a880-b27d-11e8-9f88-b639df273582.jpg)

It is all the cryptical scriblings at the bottom of the page, the so-called critical apparatus, that is the focus of this package. That apparatus of critical notes shows how a range of witnesses to a text (in my own case just two manuscripts which of course is much simpler than the hundreds of manuscripts and papyrus fragments in the Greek bible) differ and are related. All that information is very densely represented, as you can see, but it also serves as a direct tool for any domain expert to glean important information about the history of a text, relation between its witnesses and alternative readings of any passages.
As you can see, this is not just everyday pedestrial footnotes.

My package is written to make that process of establishing that critical apparatus more accurate and less error-prone. In my view it therefore contributes directly to the improvement of research, as it is used to produce research results in their own right, but those results are also the direct basis for any further research within a text discipline.

If you would like me to, I should be very happy to make these connections clear in the paper. It may be made more clear what is mean by a *critical* edition in the article. It is kept out for the sake of brevity, as (and I mean absolutely no offence by saying this) I believe that the research application is immediately obvious to to any researcher who uses the software to improve his critical textual editions.

I would also be happy to provide you with any further information about the field, if you have the stamina to evaluate the research literature.
