### 2018-09-10 #810 @Kevin-Mattheus-Moerman
@olivierverdier @JelteF @gpoore @alvinwan @maieul @adunning @kfeuerherm @pharos-alexandria @awtrindade would you be interested in reviewing this submission (_Samewords: Word disambiguation in critical text editions_) for the [Journal of Open Source Software (JOSS)](http://joss.theoj.org/about)? [The review process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#the-joss-review-process) focuses on [the software :floppy_disk:](https://github.com/stenskjaer/samewords) and [this short paper :scroll:](http://joss.theoj.org/about).
This submission features python (for critical text editing) with LaTeX, even if the actual application is not your field of expertise we would still appreciate review contributions in relation to the software itself. We aim to recruit multiple reviewers, some of whom are ideally working in a field related to this submission, however some may not work in this field but their software skills overlap such that they may judge the quality of the software.
