### 2018-09-10 #810 @Kevin-Mattheus-Moerman
@maieul thanks for your reply. You are probably hinting at a potential conflict of interest since you developed the "reledmac" package. Thanks for bringing this to our attention. Since we'll recruit multiple reviewers, and since you are not an author of the work presented here, I do not consider this to be a serious conflict of interest issue. If you are still interested I'd welcome you as a reviewer given your expertise in this field.
