### 2018-09-10 #810 @maieul
@Kevin-Mattheus-Moerman my opinion : I don't see any problem in the article, except that is may be elliptic on two points:
- what is a critical edition
- in which way editor must mark the text manually with reledmac, and what the software does could be a good point

I don't know the FOSS, but I think the article should be more explicit for not philologist reader.
