### 2018-09-10 #810 @Kevin-Mattheus-Moerman
@maieul thanks for your comments. @stenskjaer can start working on these comments. BTW as a side not this is currently a _pre-review issue_. You are welcome to start reviewing already but once I have some additional reviewers lined up I'll open the actual _review issue_. A review issue looks like [this](https://github.com/openjournals/joss-reviews/issues/845) and contains a checklist which guides reviewers through the review process. Thanks again for your help!
