### 2018-09-11 #810 @stenskjaer
@Kevin-Mattheus-Moerman This is great! Should I start revising the article based on @maieul's comments or wait until we are furhter in the review process (comments from other reviewer and you as the editor?).
