### 2018-09-11 #810 @stenskjaer
Another possible reviever could be @MarjorieBurghart who first raised the problem within the context of the reledmac LaTeX package (https://github.com/maieul/ledmac/issues/467).
