### 2018-09-12 #810 @Kevin-Mattheus-Moerman
@Padlina @MarjorieBurghart would you be interested in reviewing this submission (_Samewords: Word disambiguation in critical text editions_) for the [Journal of Open Source Software (JOSS)](http://joss.theoj.org/about)? [The review process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#the-joss-review-process) focuses on [the software :floppy_disk:](https://github.com/stenskjaer/samewords) and [this short paper :scroll:](http://joss.theoj.org/about).
