### 2018-09-12 #810 @Kevin-Mattheus-Moerman
@MarjorieBurghart @Padlina Excellent thanks! We generally aim for 2 weeks but there is not real set deadline. Some reviews take days others months but we of course prefer a speedy process.
