### 2018-09-12 #810 @Kevin-Mattheus-Moerman
@MarjorieBurghart are [this](https://www.digitalmanuscripts.eu/people/co-ordinators/marjorie-burghart/) and [this](http://ciham.ish-lyon.cnrs.fr/membres/marjorie-burghart) accurate professional links for you?  @Padlina could you share your full name and preferably a link to a professional profile? Apologies I could not tell much from your GitHub profiles.
