### 2018-09-13 #810 @gpoore
@Kevin-Mattheus-Moerman I don't have any expertise in this area, just experience with Python and LaTeX. Unfortunately, I'm currently not able to take on any additional projects until at least December, so I would recommend finding others who can review this sooner.
