### 2018-09-13 #810 @Padlina
@Kevin-Mattheus-Moerman My full name is Roberta Padlina. I don't really have a professional profile online, except for LinkedIn ([here](https://www.linkedin.com/in/roberta-padlina-6589591)) which isn't really updated. I'm a PhD student in Medieval Philosophy at the University of Fribourg, but I also work in the DH field: I worked a long time for e-codices ([my old page](https://lettres.unifr.ch/fr/hist/histoire/sciences-auxiliaires-de-lhistoire/collaborateurs/padlina.html)) and I'm working now since a year for the NIE-INE project at the University of Basel ([here](http://www.fee.unibas.ch/nie_ine.html)). Is this info enough?
