### 2018-09-13 #941 @whedon
[REVIEW]: Samewords: Word disambiguation in critical text editions
**Submitting author:** @Stenskjaer (<a href="http://orcid.org/0000-0002-8190-679X">Michael Stenskjær Christensen</a>)
**Repository:** <a href="https://github.com/stenskjaer/samewords" target ="_blank">https://github.com/stenskjaer/samewords</a>
**Version:** 0.5.6.

- [x] **Version:** 0.5.6.

---

- [x] **Version:** 0.5.6.
