### 2018-09-13 #941 @Kevin-Mattheus-Moerman
@stenskjaer, @maieul, @MarjorieBurghart, and @Padlina this is where the review happens. There are tickboxes (and information) and the top of this issue which will guide you through the review process. You may also check out our [review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). Let me know if you have any questions.
You can comment on and discuss issue here. For larger discussions/issue you are encouraged to create issues on the project repository and refer to them here.
Let the reviewing begin! :rocket:
