### 2018-10-10 #941 @Kevin-Mattheus-Moerman
Thanks @maieul. The `statement of need` is about the software of course but should be clearly present in both the paper and the documentation. If either of these do not clearly state what problems the software is designed to solve and who the target audience is, then please mention that here so the authors can work on adding these.
