### 2018-10-10 #941 @maieul
@stenskjaer

As mentionned earlier, I think that for a journal as floss, you must be more explicit about the problem your software deals, and present at least some information about what is a critical edition.
