### 2018-10-18 #941 @Padlina
@Kevin-Mattheus-Moerman Just to be sure that I'm testing everything needed: documentation is [this](https://samewords.readthedocs.io/en/latest/index.html) and software paper is [this](https://github.com/openjournals/joss-papers/blob/joss.00941/joss.00941/10.21105.joss.00941.pdf), right? Should we also test the API interface and the web service?
