### 2018-10-18 #941 @stenskjaer
@Padlina It is my understanding that you should not test the API and web service, as it is an assistive or auxilliary way of using the software. At least that is my expectation. Comments are of course welcome.
@Kevin-Mattheus-Moerman may correct me, if I am wrong.

@maieul Thank you for the observation. I think it is a good point.

Also, @Kevin-Mattheus-Moerman,  I think it is most meaningful that I make changes when all the reviews are done so I can get an overview of how to respond to the feedback. If you editors have good experience with other procedures, let me know.
