### 2018-10-18 stenskjaer/samewords#36 @Padlina
sensitive_context_match isn't false by default and has some problems
This paragraph won't get annotated without setting the sensitive_context_match parameter to false in a config file:
```
\pstart
A word with \edtext{a word}{\Afootnote{another D}} after it.
\pend
```

Moreover, once the paragraph is annotated, the resulting code is like this:
```
\pstart
\sameword{A} \sameword{word} with \edtext{\sameword[1]{a} \sameword[1]{word}}{\Afootnote{another D}} after it.
\pend
\pstart
\sameword{a} \sameword{word} with \edtext{\sameword[1]{a} \sameword[1]{word}}{\Afootnote{another D}} after it.
\pend
```

but, once compiled, the two critical notes are different:
```
6 a word2] another D
7 a2 word2 ] another D
```

And if you put the parameter multiword on true, the resulting code is this:
```
\pstart
\sameword{A word} with \edtext{\sameword[1]{a word}}{\Afootnote{another D}} after it.
\pend
\pstart
\sameword{a word} with \edtext{\sameword[1]{a word}}{\Afootnote{another D}} after it.
\pend
```

but, again, the two critical notes are different:
```
6 a word] another D
7 a word2] another D
```
