### 2018-10-18 #941 @Padlina
@stenskjaer I think that there is some issues with the sensitive context match. See stenskjaer/samewords/issues/36
