### 2018-10-18 #941 @Kevin-Mattheus-Moerman
@Padlina yes we can skip the API and web service for now. Focus on software functionality claimed/presented in the paper.

@stenskjaer in relation to waiting for all review comments to come in, I would recommend tackling issues one at a time as they come in. This way, I think, you are more likely to keep reviewers involved (we've had occasions where communication with reviewers dried up because the process took too long), and has so far been the smoothest approach. But, I leave this up to you, what ever you prefer.
