### 2018-10-18 stenskjaer/samewords#36 @stenskjaer
I suspect that the reason the case sensitivity is not enfored is because of *reledmac*. This is not clear from my documentation, but you have to enable case sensitivity in reledmac with the command `swcaseinsensitive`.

Example:
```
\documentclass{article}

\usepackage[swcaseinsensitive]{reledmac}

\begin{document}

\beginnumbering
\pstart
\sameword{A word} with \edtext{\sameword[1]{a word}}{\Afootnote{another D}} after it.
\pend
\pstart
\sameword{a word} with \edtext{\sameword[1]{a word}}{\Afootnote{another D}} after it.
\pend
\endnumbering

\end{document}
```

That gives me the app
```
1 a word2 ] another D
2 a word2 ] another D
```

The other thing about the wrong default behaviour (compared to the documentation) is a mistake.
Both these things should be adressed.
