### 2018-10-18 25c599b
Correct inaccuracy in documentation about case sensitivity setting
And add instruction to load reledmac with `swcaseinsensitive`.
