### 2018-10-18 stenskjaer/samewords#36 @stenskjaer
I have updated the documentation so that it should not reflect my original intention of having sensitive matching be the default.

The update is reflected on https://samewords.readthedocs.io/en/latest/configuration.html.
