### 2018-10-18 stenskjaer/samewords#36 @stenskjaer
Ahh -- maybe I should wait with closing the issue until you have confirmed that you think it makes sense now.
