### 2018-11-20 #941 @Kevin-Mattheus-Moerman
@maieul, @MarjorieBurghart, @Padlina can you provide an update on where we stand in the review process? Has @stenskjaer worked on the issues you highlighted? Please tick as many boxes as you are currently able to at the top. Also please list any key issues you feel are preventing you from ticking the last boxes. Thanks.
