### 2018-11-20 3512f16
Add paragraphs to the JOSS paper about its context and use
