### 2018-11-20 #941 @stenskjaer
@Kevin-Mattheus-Moerman I have added some paragraphs to give the context suggested by @maieul and pushed it to the repo of the software. But I don't know if I should trigger a rebuild of the article from in here, or somewhere else?

Further, these changes (caused by the reviews under progress) has moved the history of the repo past the 0.5.0 release flag in the repo. Should it be updated or indicated that the HEAD of the submitted material is ahead of the version indicated in the description of this review?
