### 2018-11-20 #941 @Kevin-Mattheus-Moerman
@stenskjaer thanks for the update. You can call `@whedon generate pdf` here at any time to recreate the PDF.
In relation to the version number. That is fine. Thanks for letting us know. We will make sure we update the number if this work is accepted in JOSS.
