### 2018-11-20 #941 @Kevin-Mattheus-Moerman
@maieul can you check the [updated paper](https://github.com/openjournals/joss-papers/blob/joss.00941/joss.00941/10.21105.joss.00941.pdf)? Thanks.
