### 2019-01-09 stenskjaer/samewords#36 @Padlina
First of all, Happy new yeay!
Then, yes, it make sense to me. The only thing is that I have a problem with the swcaseinsensitive option:
! LaTeX Error: Unknown option `swcaseinsensitive' for package `reledmac'.
But I assume that this is something due to my latex configuration. So, I'll close this issue.
