### 2019-01-12 #941 @Kevin-Mattheus-Moerman
@stenskjaer have you responded to @Padlina 's "small suggestions" (@Padlina can you link to where these suggestions are discussed?)?
