### 2019-02-10 #941 @Kevin-Mattheus-Moerman
> @stenskjaer see my comment above point_up , any updates on responding to these points?

Any updates on working on the points above?
