### 2019-02-10 833572c
Add try except and error reporting for invalid reledmac doc
