### 2019-02-10 a0b7f80
Update the docs to reflect suggested changes from @Padlina
- [x] Providing a minimal working .tex example would be quite useful. [Added in the docs]
- [x] https://samewords.readthedocs.io/en/latest/: in the example, the "for" word is green, but I don't see why [formatting updated]
- [x] https://s3.amazonaws.com/samewords-app/index.html give an error [update the url]
- [x] on page https://samewords.readthedocs.io/en/latest/configuration.html, the phrase "By default all characters that are not word characters, punctuation or \ { or } is considered part of a word." => it is  not really clear whic
[…]
