### 2019-02-10 #941 @stenskjaer
Thank you for your patience. I have received the following recommendations and feedback from @Padlina  (thank you very much @Padlina, by the way!)

- [x] Providing a minimal working .tex example would be quite useful. [Added in the docs. This avoids making the paper longer than most necessary (which I get is the aim of the JOSS).]
- [x] https://samewords.readthedocs.io/en/latest/: in the example, the "for" word is green, but I don't see why [formatting updated]
- [x] https://s3.amazonaws.com/samewords-app/index.html give an error [update the url]
- [x] on page https://samewords.readthedocs.io/en/latest/configuration.html, the phrase "By default all characters that are not word characters, punctuation or \ { or } is considered part of a word." => it is  not really clear which special characters are also considered part of a word. [Phrasing modified]
- [x] In the same page, "If you feel bold you could of course edit the punctuation list in the settings file". => specify which settings file [That sentence is not needed and has been removed]
- [x] I would provide an example of configuration settings for each parameter [I have added a reference to where the example configuration file can be found (in the root directory of the package and the online github repo with a link to https://github.com/stenskjaer/samewords/blob/master/sample_config.json)]
- [x] Under the section "multiword", the third paragraph ("A word with edtext{a word}{Afootnote{another D}} after it.") should be displayed as code [Actually it is all a text example to it is now a proper code block.]
- [x] Section "sensitive_context_match": "The value of the settings variable sensitive_proximity_match" => which one? sensitive_context_match or sensitive_proximity_match ? [Fixed, it should be ``sensitive_context_match``]

I have pushed the updates to Github and changed the docs as described in the list above.
