### 2019-03-01 #941 @Kevin-Mattheus-Moerman
I think it is okay to skip that DOI. @openjournals/joss-eics should @maieul list the package differently in the .bib file to avoid the missing DOI warning?
