### 2019-03-01 #941 @danielskatz
The missing DOI warning is just a warning.  Just ignore it if it doesn't make sense.
