### 2019-03-04 940b539
Correct bug in documentation MWE
Closes #38
