### 2019-03-05 #941 @Kevin-Mattheus-Moerman
Great. It looks like we are in good shape.

@stenskjaer can you please do the following:

- [ ] Create an archived version of the software on a service like [ZENODO](https://zenodo.org/) (some have found [these steps](https://guides.github.com/activities/citable-code/) useful for automated archival of releases), and please list the DOI here once completed?

- [ ] Can you tell us what the latest version (or release) tag is for the reviewed and archived software (it likely moved on from 0.5.0)?

Thanks.
