### 2019-03-26 #941 @stenskjaer
Sorry for the wait!

- The Zenodo version is at https://zenodo.org/record/2609228 (DOI: 10.5281/zenodo.2609228)
- The version is 0.5.6.
