### 2019-04-09 #941 @labarba
@stenskjaer — Since we will have the Zenodo archive DOI on the front page of the paper, as part of the metadata shown in the margin note, we don't include a citation to it or mention it in the paper. Can you delete that?
