### 2019-04-13 810e5d9
Remove references to Zenodo storage
