+++ paper.bib
@@ -8,11 +8,3 @@
     Url = {https://www.ctan.org/pkg/reledmac},
     }
 
-@misc{zenodo,
-  author       = {Michael Stenskjær Christensen},
-  title        = {{stenskjaer/samewords: Word disambiguation in critical text editions}},
-  month        = jul,
-  year         = 2018,
-  doi          = {10.5281/zenodo.1306293},
-  url          = {https://doi.org/10.5281/zenodo.1306293}
-}
