### 2019-04-13 #941 @labarba
Congratulations, @stenskjaer, your JOSS paper is published!

Big thanks to our editor: @Kevin-Mattheus-Moerman, and the reviewers: @maieul, @Padlina — we couldn't do this without you! 🙏
